chore(tooling): add "do not edit" header to generated schemas
The schema generator script has been updated to prepend a warning comment to the `genkit_schemas.dart` file.

This header clarifies that the file is auto-generated and should not be modified directly. This prevents developers from making manual edits that would be overwritten the next time the generation script is run.