refactor(stream)!: Replace FlowStreamResponse with ActionStream (#5)
* refactor(stream)!: Replace FlowStreamResponse with ActionStream

This commit introduces a breaking change to the streaming API to simplify its usage.

The `.stream()` method on a `RemoteAction` now returns a `ActionStream` object directly, instead of a `FlowStreamResponse` record.

`ActionStream` is a `Stream` subclass, so it can be used directly in `await for` loops. It also provides a `result` property to access the flow's final, non-streamed response, replacing the previous `response` future from the record. Also add `onResult` property which is a Future that completes when the final response is recieved.

This change makes the API more intuitive and reduces boilerplate when handling streaming flows.

BREAKING CHANGE: The `.stream()` method now returns an `ActionStream` instead of a `FlowStreamResponse` record (`({Future<O?> response, Stream<S> stream})`).

Migration:
- Replace `final (:stream, :response) = myAction.stream(...)` with `final stream = myAction.stream(...)`.
- Access the final result via `stream.result` instead of `await response`. Or `await stream.onResult`.